requirements.txt
protobuf
python-snappy 
polars
